Continued Working on Buff Boon System.
NOTE: If someone knows how to do all this with delagates then feel free to let me know and do that. I tried to use them but at least the way I figured out how to do them (Since Im not very familiar with delegates) it was almost the same as the way I am doing it now.